--- requirements/requirements.txt
@@ -1,2 +1,0 @@
-fastapi==0.100.0
-uvicorn==0.23.0
