--- requirements/requirements_dev.txt
@@ -1,20 +1,0 @@
-black==23.1.0
-mypy==1.4.1
-tryceratops==1.1.0
-flake8-absolute-import==1.0.0.1
-flake8-annotations==3.0.0
-flake8-async==22.11.14
-flake8-bugbear==23.2.13
-flake8-comprehensions==3.10.1
-flake8-isort==6.0.0
-pyflakes==3.0.1
-flake8-plugin-utils==1.3.2
-isort==5.12.0
-flake8-print==5.0.0
-pre-commit==3.2.0
-pycodestyle==2.10.0 
-flake8-pytest-style==1.7.2
-flake8-quotes==3.3.2
-flake8-return==1.2.0
-flake8-simplify==0.19.3
-flake8-variables-names==0.0.5
