--- requirements/requirements_test.txt
@@ -1,1 +1,0 @@
-pytest==7.4.0
